[tin d0ec1ce7] I've noticed an error in thread handling.
I often run tin commit in the middle of in-process threads, then run tin push, and then continue working on them after that. When I tin commit _again_ (and fully complete the thread), and then push the changes to the remote, the thread content isn't updated. It only shows the original thread length (the messages that were committed and pushed with the first commit).

However, the local `tin thread show` shows the full thread.

Can the thread content be updated in subsequent pushes as well?